% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aictab.R
\name{aictab}
\alias{aictab}
\title{Show a table of AIC model comparisons}
\usage{
aictab(x, plot = FALSE, models = names(x$models)[!names(x$models) \%in\%
  c("absdiff", "absunc")])
}
\arguments{
\item{x}{An RSA object}

\item{plot}{Should a plot of the AICc table be plotted? (Experimental)}

\item{models}{A vector with all model names of the candidate set. Defaults to all polynomial models in the RSA object.}
}
\description{
Show a table of AIC model comparisons
}
\details{
TODO
}

