\name{CI.boot}
\alias{CI.boot}
\title{Compute bootstrapped confidence interval and p value}
\usage{
  CI.boot(r.boot, CI = 0.95)
}
\arguments{
  \item{r.boot}{Results object from the
  \code{\link{bootRSA}} function}

  \item{CI}{Width of confidence interval}
}
\description{
  Compute bootstrapped confidence interval and p value
}
\details{
  See \code{\link{bootRSA}} for more details and example.
}
\seealso{
  \code{\link{RSA}}
}

