% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cubic.R
\name{clRange}
\alias{clRange}
\title{Range check for the CL/RRCL model}
\usage{
clRange(object, alpha = 0.05, verbose = TRUE, model = "CL")
}
\arguments{
\item{object}{An RSA object}

\item{alpha}{Alpha level}

\item{verbose}{Should extra information be printed?}

\item{model}{Either "CL" or "RRCL"}
}
\description{
Compute the regions of significance and test their intersection with the data
}
\details{
TODO
}
