% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsb_colours_fonts.R
\name{ddsb_colours_fonts}
\alias{ddsb_colours_fonts}
\title{DDSB colours and fonts}
\usage{
ddsb_colours_fonts(...)
}
\arguments{
\item{...}{No parameters}
}
\value{
R objects with names and HEX colours for DDSB visual identity
}
\description{
DDSB colours and fonts
}
\examples{
ddsb_colours_fonts()
}
