\docType{data}
\name{motcon}
\alias{motcon}
\title{Data set on motive congruence.}
\format{A data frame with 361 rows and 3 variables}
\description{
  A dataset containing the explicit intimacy motive,
  implicit affiliation motive (nAff) and negative
  activation. The variables are as follows:
}
\details{
  \itemize{ \item EM Explicit intimacy motive, measured
  with questionnaire \item IM Implicit affiliation motive,
  measure with picture story exercise. Raw motive scores
  have been controled for word count.  \item negAct
  Negative activation, measure with the NA subscale of the
  PANAVA inventory. }
}
\keyword{datasets}

