% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generics.R
\name{residuals.RSA}
\alias{resid}
\alias{residuals.RSA}
\title{Return residual values of a RSA model}
\usage{
\method{residuals}{RSA}(object, ..., model = "full")
}
\arguments{
\item{object}{An RSA object.}

\item{...}{Other parameters (currently not used)}

\item{model}{Model on which the fitted values are based}
}
\description{
Return residual values of a RSA model
}

