% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cubic.R
\name{caRange}
\alias{caRange}
\title{Range check for the CA/RRCA model}
\usage{
caRange(object, alpha = 0.05, verbose = TRUE, model = "CA")
}
\arguments{
\item{object}{An RSA object}

\item{alpha}{Alpha level}

\item{verbose}{Should extra information be printed?}

\item{model}{Either "CA" or "RRCA"}
}
\description{
Main function to identify data points behind E2 and test how many of them have outcome predictions that significantly differ from predictions for predictor combinations on E2 (that have the same level)
}
\details{
TODO
}
