\name{compare}
\alias{compare}
\title{Compare RSA models}
\usage{
compare(x, verbose = TRUE, plot = FALSE)
}
\arguments{
  \item{x}{An RSA object}

  \item{verbose}{Should the summary be printed?}

  \item{plot}{Should the comparison be plotted (using the
  \code{\link{modeltree}} function)?}
}
\description{
Compare several fit indexes of all models computed from the
RSA function
}
\details{
No details so far.
}

