% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compare}
\alias{compare}
\title{Compare a full list of RSA models}
\usage{
compare(x, verbose = TRUE, plot = FALSE, ...)
}
\arguments{
\item{x}{An RSA object}

\item{verbose}{Should the summary be printed?}

\item{plot}{Should the comparison be plotted (using the \code{\link{modeltree}} function)?}

\item{...}{Additional parameters passed to the \code{\link{modeltree}} function}
}
\description{
Compare several fit indexes of all models computed from the RSA function
}
\details{
No details so far.
}

