% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare a full list of RSA models}
\usage{
compare(x, verbose = TRUE, plot = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{An RSA object}

\item{verbose}{Should the summary be printed?}

\item{plot}{Should the comparison be plotted (using the \code{\link{modeltree}} function)?}

\item{digits}{Digits of the output}

\item{...}{Additional parameters passed to the \code{\link{modeltree}} function}
}
\description{
Compare several fit indexes of all models computed from the RSA function
}
\details{
No details so far.
}
